Fix Column 'ID' in on clause is ambiguous
The join clause introduced by Polylang could conflict with other plugins
if another join statement introduced another table with a 'ID' column.

This commit may break 3rd party plugins making use of this internal
function and taking profit of this ambiguity. This is the case of
Polylang for WooCommerce up to v0.2.1.

Fixes #4