Fix draft created when creating a new page on multisite
This reverts commit 812d03ec3de820cfb796992bc33ac3a3eda30730 and commit
67c5c47127ef24cca477f5aa015af434d8612be5
This commits were added to work around a bug in Gutenberg when js
scripts
are concatenated. This however proved to be not sufficient as a new
version of Gutenberg broke our js again and we finally decided to
develop the Gutenberg compatibility in a separate plugin.
Fix https://github.com/polylang/polylang-pro/issues/99