Move the js of the internals of the legacy metabox to its own file
This is necessary as the dependancy on post.js of the classic editor
script (tagBox) conflicts with the block editor.
We however need some part of the script (now moved to metabox.js) to
handle the automplete and internal buttons of the legacy metabox in the
block editor.
Fix https://github.com/polylang/polylang-pro/issues/184